For compatibility with float timestamps in python-2.5, convert st_mtime attributes to long wherever portage expects 1s resolution.
svn path=/main/trunk/; revision=4500

+++ pym/cache/flat_list.py
@@ -35,7 +35,8 @@ class database(fs_template.FsBased):
 				raise KeyError(cpv)
 			raise cache_errors.CacheCorruption(cpv, e)
 
-		try:	d["_mtime_"] = os.fstat(myf.fileno()).st_mtime
+		try:
+			d["_mtime_"] = long(os.fstat(myf.fileno()).st_mtime)
 		except OSError, e:	
 			myf.close()
 			raise cache_errors.CacheCorruption(cpv, e)
